+++ debugdrawer-bridge-noop/src/main/kotlin/drivechain/bridge/debugdrawer/DebugDrawerBridgeModule.kt
@@ -1,0 +1,8 @@
+import dagger.Module
+import dagger.Provides
+import io.palaima.debugdrawer.base.DebugModule
+import javax.inject.Singleton
+
+@Module class DebugDrawerBridgeModule {
+    @Provides @Singleton fun debugModuleSet() : Set<DebugModule> = setOf() // empty set for release
+}