+++ debugdrawer-logs-bridge/src/main/kotlin/drivechain/bridge/debugdrawer/DebugDrawerLogsBridgeModule.kt
@@ -1,0 +1,11 @@
+package drivechain.bridge.debugdrawer
+
+import dagger.Module
+import dagger.Provides
+import dagger.multibindings.IntoSet
+import io.palaima.debugdrawer.base.DebugModule
+import io.palaima.debugdrawer.logs.LogsModule
+
+@Module class DebugDrawerLogsBridgeModule {
+    @Provides @IntoSet fun logsModule() : DebugModule = LogsModule()
+}